+++ src/feature/hs/hs_intropoint.c
@@ -205,7 +205,7 @@ handle_verified_establish_intro_cell(or_circuit_t *circ,
   circuit_change_purpose(TO_CIRCUIT(circ), CIRCUIT_PURPOSE_INTRO_POINT);
   /* Initialize the INTRODUCE2 token bucket for the rate limiting. */
   token_bucket_ctr_init(&circ->introduce2_bucket, hs_dos_get_intro2_rate(),
-                        hs_dos_get_intro2_burst(), approx_time());
+                        hs_dos_get_intro2_burst(), (uint32_t) approx_time());
 
   return 0;
 }
